+++ src/styles/base/colors.js
@@ -1,0 +1,32 @@
+// Color palette
+const [
+  grey,
+  black, // eslint-disable-line no-unused-vars
+  white, // eslint-disable-line no-unused-vars
+  red, // eslint-disable-line no-unused-vars
+  orange, // eslint-disable-line no-unused-vars
+  yellow,
+  green, // eslint-disable-line no-unused-vars
+  blue, // eslint-disable-line no-unused-vars
+  indigo, // eslint-disable-line no-unused-vars
+  violet // eslint-disable-line no-unused-vars
+] = [
+  '#333',
+  '#111',
+  '#fff',
+  '#e74c3c',
+  '#e67e22',
+  '#f1c40f',
+  '#2ecc71',
+  '#3498db',
+  '#34495e',
+  '#9b59b6'
+]
+
+const primaryColor = yellow
+const secondaryColor = grey
+
+export default {
+  primaryColor,
+  secondaryColor
+}