+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.0.2 - 2017-03-05
+### Added
+* aphrodite
+* css-loader
+* style-loader
+* HMR updates
+* eslint-loader
+
 ## 0.0.1 - 2017-03-03
 ### Added
-- Initial creation
+* Initial creation